combine batch and transformer engine
